Dependencies switched to NuGet; minor ver updates
* All project-to-project references switched to NuGet references except
in the case of unit tests referencing exactly the project they test.
This includes the example projects, which ONLY refer to NuGet now. The
examples must refer to NuGet because the versions in the AssemblyInfo.cs
files don't have the published versions from PackageVersions.proj.
* Autofac.Mvc5.Owin nuspec updated to remove dependency on Autofac.Mvc5
(dependency not required).
* Autofac.Extras.Attributed nuspec updated to require a dependency on
Autofac.Integration.Mef (required but previously not specified).
* Autofac.Extras.DomainServices nuspec updated to require a later
version of Autofac (3.4.0) and require Autofac.Web (3.2.0). Autofac.Web
was required before but not specified by the nuspec.
* Autofac.Extras.EnterpriseLibraryConfigurator nuspec updated to remove
dependency on Autofac.Extras.CommonServiceLocator (dependency not
required).
* Autofac.Mvc5.Owin version updated to 3.0.0-alpha4 to reflect removal
of the dependency.
* Autofac.Extras.Attributed version updated to 3.3.0 to reflect the new
dependency requirement.
* Autofac.Extras.DomainServices version updated to 3.1.0 to reflect the
new dependency requirement.
* Autofac.Extras.EnterpriseLibraryConfigurator version updated to 3.0.2
to reflect the removal of the dependency.